Move object definition type
