+++ types/Schemas.ts
@@ -62,3 +62,27 @@ export type SchemaDefinition = {
   ];
   requiredProperties: Array<string>;
 };
+
+export type ObjectDefinition = {
+  inputs: [
+    {
+      associations?: [
+        {
+          types: [
+            {
+              associationCategory: string;
+              associationTypeId: number;
+            },
+          ];
+          to: {
+            id: string;
+          };
+        },
+      ];
+      objectWriteTraceId?: number;
+      properties: {
+        [key: string]: string;
+      };
+    },
+  ];
+};
